v2023.5 (#30)
### Physically significant changes:

- deltaV (cell volume) was being calculated with the expansion time at
the current timestep (nts) instead of nts_prev. This has been reverted
to use ntsprev, which matches ARTIS classic. The smaller deltaV will
slightly increase normalised values of estimators such as J, nuJ, etc.

- For calculating grey depth of a cell (to decide if it should be
treated in thick/grey mode), the distance to observer is now clamped to
zero. The observer is assumed to be at radial distance xmax, which leads
to negative distances in the cube corners in 3D mode. Setting cells with
grey depth >= 0 would have previously excluded cells in the cube corners
(with negative optical depths).

- if USE_RELATIVISTIC_DOPPLER_SHIFT is on, gamma-ray heating
rpkt_emiss[mgi] is now calculated with relativistic correction, which
fixes an issue with negative gamma heating rates for high velocity cells
(>~0.5c).

### New features

- an artis.pid file keeps track of the (rank 0) process id and prevents
two copies of artis (or artis and exspec) from running simultaneously in
the same directory.
- 3D model.txt files can be read in either one line (previous) or two
line per cell format. The one line per cell format makes file reading
and parsing by standard tools easier. artistools decompresses and reads
these files twice to get even and odd rows, which is slow if xzip is
used on a large model file.
- New option USE_GSL_RANDOM can be set to false to use std::random in
place of GSL's random number generation. So far, this appears to be a
slower method, so it has been switched off.
- Both phixsdata.txt and phixsdata_v2.txt will be read if they both
exist. (Was also added to v2022.11.1)
- New option CLASSIC_LINELIST_SORT can be set to false (default for
kilonova options). The classic line sort function clamps close lines
together and does some memory unsafe things that can result in a
linelist that is not strictly sorted by frequency.
- a new classic 3D test runs a 10^3 model with classic options.
- the classic test now uses a classic atomic dataset (phixsdata.txt)

### Performance improvements
- 1D Sim+2010 subMch model in classic mode with 2e7 packets on 960 cores
(JUWELS) completed 16% faster than v2022.11. Better agreement with
classic, mostly due to the deltaV at nts_prev reversion.
- Reduced lookups of nucindex in decay routines speed up initial
temperature estimation and estimator writing.
- Duplication of atomic transitions in element/ion/level order increases
performance when calculating cooling rates and the macroatom.
- Kilonova models run about 100% faster.
- Use of cumulative sums for macroatom transition probability lists, and
then binary searches speed up do_macroatom, which dominates the runtime
for kilonova models.

### Removals
- The unused packet property scat_count was removed.
- Behaviour according to do_comp_est = false, do_r_lc = true, do_rlc_est
= 3 is always on, and these variables have now been removed.
- SF_USE_LOG_E_INCREMENT. The experimental log energy increment made the
Spencer-Fano solution less accurate and was never used in production
runs.
- DETAILED_BF_ESTIMATORS_BYTYPE. This experimental code was extremely
fragile and memory consuming and did not work in MPI mode.
- NT_BLOCKSIZEEXCITATION. std::vector now handles the dynamic allocation
of the excitation list.
- MINDENSITY. This option didn't actually do anything.
### Renames
- artisoptions: Replace NO_INITIAL_PACKETS with INITIAL_PACKETS_ON
- artisoptions: Replace NO_LUT_PHOTOION with USE_LUT_PHOTOION
- artisoptions: Replace NO_LUT_BFHEATING with USE_LUT_BFHEATING

+++ scripts/exspec-before.sh
@@ -1,12 +1,23 @@
 #!/usr/bin/env bash
 
-# files outside the artis folder, so move them back
-if [ ! -f packets00_0000.out* ]; then
+# packet output files outside the artis folder, so move them back to run exspec
+if [[ ! -f packets00_0000.out* && -f packets/packets00_0000.out* ]]; then
   mv packets/packets*.out* .
 fi
 
-gzip -d packets*.out.gz || true
-xz -d -T 0 packets*.out.xz || true
+find . -maxdepth 1 -name 'packets**.out.zst' -exec zst -d -v -T0 --rm {} \;
+find . -maxdepth 1 -name 'packets**.out.lz4' -exec lz4 -d -v --rm {} \;
+find . -maxdepth 1 -name 'packets**.out.gz' -exec gzip -d -v {} \;
+find . -maxdepth 1 -name 'packets**.out.xz' -exec xz -d -v -T0 {} \;
 
-gzip -d -v phixsdata_v2.txt.gz transitiondata.txt.gz ratecoeff.dat.gz || true
-xz -d -v -T 0 phixsdata_v2.txt.xz transitiondata.txt.xz ratecoeff.dat.xz || true
+#unzip, e.g. phixsdata_v2.txt.xz transitiondata.txt.xz ratecoeff.dat.xz
+
+find . -maxdepth 1 -name '*.dat.zst' -exec zst -d -v -T0 --rm {} \;
+find . -maxdepth 1 -name '*.dat.lz4' -exec lz4 -d -v --rm {} \;
+find . -maxdepth 1 -name '*.dat.xz' -exec xz -d -v -T0 {} \;
+find . -maxdepth 1 -name '*.dat.gz' -exec gzip -d -v {} \;
+
+find . -maxdepth 1 -name '*.txt.zst' -exec zst -d -v -T0 --rm {} \;
+find . -maxdepth 1 -name '*.txt.lz4' -exec lz4 -d -v --rm {} \;
+find . -maxdepth 1 -name '*.txt.xz' -exec xz -d -v -T0 {} \;
+find . -maxdepth 1 -name '*.txt.gz' -exec gzip -d -v {} \;
